ui: extract dataFromServer global from login file
Since the dataFromServer will hold more than just login state, move it out to a
new file.

Release note: None
