+++ pkg/ui/src/util/dataFromServer.ts
@@ -1,0 +1,18 @@
+export interface DataFromServer {
+  ExperimentalUseLogin: boolean;
+  LoginEnabled: boolean;
+  LoggedInUser: string;
+}
+
+// Tell TypeScript about `window.dataFromServer`, which is set in a script
+// tag in index.html, the contents of which are generated in a Go template
+// server-side.
+declare global {
+  interface Window {
+    dataFromServer: DataFromServer;
+  }
+}
+
+export function getDataFromServer(): DataFromServer {
+  return window.dataFromServer || {};
+}
